pyproject.toml
[project]
name = "billing-rest-api"
version = "0.1.0"
description = "REST API для управления пользователями, счетами и платежами"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.0",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pwdlib[argon2]>=0.2.1",
    "psycopg[binary]>=3.2.9",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "testcontainers[postgres]>=4.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
